Don't expose NonZero through libstd.
